+++ indico/modules/events/persons/controllers.py
@@ -13,7 +13,9 @@ from marshmallow import fields
 from sqlalchemy import and_, or_
 from sqlalchemy.orm import contains_eager, joinedload
 from webargs import validate
+from werkzeug.exceptions import Forbidden
 
+from indico.core.config import config
 from indico.core.db import db
 from indico.core.db.sqlalchemy.custom.unaccent import unaccent_match
 from indico.core.db.sqlalchemy.principals import PrincipalType
@@ -381,6 +383,11 @@ class RHEditEventPerson(RHPersonsBase):
 
 
 class RHEventPersonSearch(RHAuthenticatedEventBase):
+    def _check_access(self):
+        RHAuthenticatedEventBase._check_access(self)
+        if not config.ALLOW_PUBLIC_USER_SEARCH and not self.event.can_manage(session.user):
+            raise Forbidden
+
     def _search_event_persons(self, exact=False, **criteria):
         criteria = {key: v for key, value in criteria.items() if (v := value.strip())}