+++ indico/modules/users/client/js/Favorites.jsx
@@ -21,14 +21,17 @@ import {indicoAxios, handleAxiosError} from 'indico/utils/axios';
 
 import './Favorites.module.scss';
 
-window.setupFavoriteSelection = function setupFavoriteSelection(userId) {
-  ReactDOM.render(<FavoriteManager userId={userId} />, document.getElementById('user-favorites'));
+window.setupFavoriteSelection = function setupFavoriteSelection(userId, userSearchDisabled) {
+  ReactDOM.render(
+    <FavoriteManager userId={userId} userSearchDisabled={userSearchDisabled} />,
+    document.getElementById('user-favorites')
+  );
 };
 
-function FavoriteManager({userId}) {
+function FavoriteManager({userId, userSearchDisabled}) {
   return (
     <>
-      <FavoriteUserManager userId={userId} />
+      <FavoriteUserManager userId={userId} userSearchDisabled={userSearchDisabled} />
       <FavoriteCatManager userId={userId} />
     </>
   );
@@ -36,10 +39,12 @@ function FavoriteManager({userId}) {
 
 FavoriteManager.propTypes = {
   userId: PropTypes.number,
+  userSearchDisabled: PropTypes.bool,
 };
 
 FavoriteManager.defaultProps = {
   userId: null,
+  userSearchDisabled: false,
 };
 
 function FavoriteCatManager({userId}) {
@@ -135,7 +140,7 @@ FavoriteCatManager.defaultProps = {
   userId: null,
 };
 
-function FavoriteUserManager({userId}) {
+function FavoriteUserManager({userId, userSearchDisabled}) {
   const [favoriteUsers, [addFavoriteUser, deleteFavoriteUser], loading] = useFavoriteUsers(userId);
 
   const searchTrigger = triggerProps => (
@@ -189,19 +194,23 @@ function FavoriteUserManager({userId}) {
           )}
         </div>
       </div>
-      <UserSearch
-        existing={Object.values(favoriteUsers).map(u => u.identifier)}
-        onAddItems={e => e.forEach(u => addFavoriteUser(u.userId))}
-        triggerFactory={searchTrigger}
-      />
+      {!userSearchDisabled && (
+        <UserSearch
+          existing={Object.values(favoriteUsers).map(u => u.identifier)}
+          onAddItems={e => e.forEach(u => addFavoriteUser(u.userId))}
+          triggerFactory={searchTrigger}
+        />
+      )}
     </div>
   );
 }
 
 FavoriteUserManager.propTypes = {
   userId: PropTypes.number,
+  userSearchDisabled: PropTypes.bool,
 };
 
 FavoriteUserManager.defaultProps = {
   userId: null,
+  userSearchDisabled: false,
 };